Split integration tests into two files
